PAINTROID-24: Refactor fill tool
* Move fill algorithm tests to `FillAlgorithmTest`
* Move fill command tests to `FillCommandTest`
* Refactor `FillToolTest` to work without an activity
* Refactor all UI dependencies out of `FillTool` into `DefaultFillToolOptions`
